+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 -
 
 ## Misc:
-- 
+- Improved file checks by automating some exclusions and replacements through regexes
 
 # 1.4.0
 ## Bug Fixes: